Moved a duplicated test to common and deleted its counterpart; in doing so found some places in common that still had DB-specific references, so removed those.
